% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pct.R
\name{calc_pct}
\alias{calc_pct}
\title{Calculate percentile values and percentile plot based on tabulated normative mean and SD}
\usage{
calc_pct(
  mean,
  sd,
  agecat,
  data = NULL,
  agecatData = NULL,
  sample = NULL,
  description = NULL,
  title = NULL,
  linesize = 1.5,
  showData = FALSE
)
}
\arguments{
\item{mean}{A numeric value representing the normative mean values}

\item{sd}{A numeric value representing the normative standard deviation values}

\item{agecat}{A factor variable representing the age category of the normative values}

\item{data}{A data frame representing the sample data}

\item{agecatData}{A factor variable representing the age categories of each sample}

\item{sample}{A numeric vector representing the collected measurements}

\item{description}{A string representing the description of the normative values}

\item{title}{A string representing the plot title}

\item{linesize}{A numeric value representing the line size of the percentile curves (default = 1.5)}

\item{showData}{A logical value indicating whether to show data in the plot (default =FALSE)}
}
\value{
A list with normative values and a percentile plot
}
\description{
Calculate percentile values and percentile plot based on tabulated normative mean and SD
}
